:white_check_mark: Update tests - CommitMessage type changed
